+++ packages/sources/src/components/form/form.models.ts
@@ -4,6 +4,7 @@ export * from './models/form-group-files.model';
 export * from './models/form-group-input.model';
 export * from './models/form-group-input-number.model';
 export * from './models/form-group-radios.model';
+export * from './models/form-group-search-bar.model';
 export * from './models/form-group-select.model';
 export * from './models/form-group-static.model';
 export * from './models/form-group-textarea.model';